fix(tests): Remove the Docker containers after test execution
Use linktime crate to setup constructor/destructors for global
setup/teardown in the IT tests.

Using rstest's `[once]` didn't work because ConnectionManager's Drop was
not called.

Export `DF_TABLE_PROVIDERS_DEBUG` env var with any value if the Docker
container should be stopped, but not removed. For debugging purposes.

Use u16 for the `port` variable

+++ core/tests/mongodb/mod.rs
@@ -16,7 +16,7 @@ use crate::docker::RunningContainer;
 
 mod common;
 
-async fn test_mongodb_datetime_types(port: usize) {
+async fn test_mongodb_datetime_types(port: u16) {
     let ts0 = DateTime::parse_from_rfc3339("2024-09-12T10:00:00Z")
         .unwrap()
         .with_timezone(&Utc);
@@ -100,7 +100,7 @@ async fn test_mongodb_datetime_types(port: usize) {
 }
 
 #[allow(clippy::approx_constant)]
-async fn test_mongodb_numeric_types(port: usize) {
+async fn test_mongodb_numeric_types(port: u16) {
     let decimal = Decimal128::from_str("123.456").unwrap();
 
     let test_docs = vec![doc! {
@@ -141,7 +141,7 @@ async fn test_mongodb_numeric_types(port: usize) {
     arrow_mongodb_one_way(port, "numeric_collection", test_docs, expected_record, None).await;
 }
 
-async fn test_mongodb_string_types(port: usize) {
+async fn test_mongodb_string_types(port: u16) {
     let test_docs = vec![
         doc! {
             "name": "Alice",
@@ -177,7 +177,7 @@ async fn test_mongodb_string_types(port: usize) {
     arrow_mongodb_one_way(port, "string_collection", test_docs, expected_record, None).await;
 }
 
-async fn test_mongodb_boolean_types(port: usize) {
+async fn test_mongodb_boolean_types(port: u16) {
     let test_docs = vec![
         doc! {
             "is_active": true,
@@ -210,7 +210,7 @@ async fn test_mongodb_boolean_types(port: usize) {
     arrow_mongodb_one_way(port, "boolean_collection", test_docs, expected_record, None).await;
 }
 
-async fn test_mongodb_binary_types(port: usize) {
+async fn test_mongodb_binary_types(port: u16) {
     let test_docs = vec![doc! {
         "binary_data": Bson::Binary(mongodb::bson::Binary {
             subtype: mongodb::bson::spec::BinarySubtype::Generic,
@@ -239,7 +239,7 @@ async fn test_mongodb_binary_types(port: usize) {
     arrow_mongodb_one_way(port, "binary_collection", test_docs, expected_record, None).await;
 }
 
-async fn test_mongodb_object_id_types(port: usize) {
+async fn test_mongodb_object_id_types(port: u16) {
     let oid1 = mongodb::bson::oid::ObjectId::new();
     let oid2 = mongodb::bson::oid::ObjectId::new();
 
@@ -273,7 +273,7 @@ async fn test_mongodb_object_id_types(port: usize) {
 }
 
 #[allow(clippy::approx_constant)]
-async fn test_mongodb_array_types(port: usize) {
+async fn test_mongodb_array_types(port: u16) {
     let test_docs = vec![
         doc! {
             "string_tags": ["rust", "mongodb", "arrow"],
@@ -390,7 +390,7 @@ async fn test_mongodb_array_types(port: usize) {
     arrow_mongodb_one_way(port, "array_collection", test_docs, expected_record, None).await;
 }
 
-async fn test_mongodb_nested_object_types(port: usize) {
+async fn test_mongodb_nested_object_types(port: u16) {
     let test_docs = vec![
         doc! {
             "user": {
@@ -564,7 +564,7 @@ async fn test_mongodb_nested_object_types(port: usize) {
     assert_eq!(string_array.value(1), "also not an object");
 }
 
-async fn test_mongodb_null_and_missing_fields(port: usize) {
+async fn test_mongodb_null_and_missing_fields(port: u16) {
     let test_docs = vec![
         doc! {
             "name": "Alice",
@@ -616,7 +616,7 @@ async fn test_mongodb_null_and_missing_fields(port: usize) {
 }
 
 async fn arrow_mongodb_one_way(
-    port: usize,
+    port: u16,
     collection_name: &str,
     test_docs: Vec<Document>,
     expected_record: RecordBatch,
@@ -692,7 +692,7 @@ async fn arrow_mongodb_one_way(
     record_batches
 }
 
-async fn test_mongodb_unnesting_depth_1(port: usize) {
+async fn test_mongodb_unnesting_depth_1(port: u16) {
     let ts0 = DateTime::parse_from_rfc3339("2024-09-12T10:00:00Z")
         .unwrap()
         .with_timezone(&Utc);
@@ -773,7 +773,7 @@ async fn test_mongodb_unnesting_depth_1(port: usize) {
 /// top-level while every other document field — scalar, nested document, and
 /// array — folds into one sorted-key JSON `Utf8` catch-all column (`data`).
 /// Exercised end-to-end through the DataFusion scan path against a live MongoDB.
-async fn test_mongodb_json_nesting(port: usize) {
+async fn test_mongodb_json_nesting(port: u16) {
     use datafusion_table_providers::schema_projection::SchemaProjection;
 
     let test_docs = vec![
@@ -906,7 +906,7 @@ fn project_record_batch(batch: &RecordBatch, columns: &[&str]) -> DFResult<Recor
         .map_err(|e| DataFusionError::ArrowError(Box::new(e), None))
 }
 
-async fn start_mongodb_container(port: usize) -> RunningContainer {
+async fn start_mongodb_container(port: u16) -> RunningContainer {
     let running_container = common::start_mongodb_docker_container(port)
         .await
         .expect("MongoDB container to start");
@@ -945,7 +945,7 @@ async fn test_mongodb_arrow_oneway() {
 /// `MongoDBExec::try_pushdown_sort` returns `Exact` — silently losing the N.
 /// These tests verify end-to-end that `SELECT ... ORDER BY ... LIMIT N`
 /// returns exactly N rows.
-async fn test_mongodb_sort_limit(port: usize) {
+async fn test_mongodb_sort_limit(port: u16) {
     let ctx = SessionContext::new();
     let client = common::get_mongodb_client(port)
         .await
